// src/types.ts
export interface AuthAccountEntry {
	numberOfSignatures: number;
	mandatoryKeys: string[];
	optionalKeys: string[];
	nonce: string;
}

export interface AuthStoreEntry {
	address: string;
	authAccount: AuthAccountEntry;
}
// ...
export interface SupplySubstoreEntry {
	tokenID: string;
	totalSupply: string;
}

export interface EscrowSubstoreEntry {
	escrowChainID: Buffer;
	tokenID: Buffer;
	amount: bigint;
}

export interface SupportedTokensSubstoreEntry {
	chainID: Buffer;
	supportedTokenIDs: Buffer[];
}
// ...
export interface GenesisDataEntry {
	initRounds: number;
	initValidators: string[];
}

export interface PoSStoreEntry {
	validators: ValidatorEntry[];
	stakers: Staker[];
	genesisData: GenesisDataEntry;
}

export interface GenesisAssetEntry {
	module: string;
	data: Record<string, unknown>;
	schema: Schema;
}
// ...
export interface NetworkConfigLocal {
	name: string;
	tokenID: string;
	prevSnapshotBlockHeight: number;
	additionalAccounts: { address: Buffer; balance: bigint }[];
}